pyproject.toml
[project]
name = "conceptor-python-scripts"
version = "0.1.0"
description = ""
authors = [
    {name = "Ed Ropple",email = "ed@edropple.com"}
]
license = {text = "UNLICENSED"}
package-mode = false
requires-python = "3.11.*"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
